% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prmun}
\alias{prmun}
\title{prmun}
\format{
A data frame with 399 rows and 4 variables:
\describe{
\item{\code{CD_MUN}}{character código do município}
\item{\code{NM_MUN}}{character nome do município}
\item{\code{SIGLA_UF}}{character sigla da unidade da federação}
\item{\code{AREA_KM2}}{double área em KM2}
}
}
\source{
\url{http://github.com/abreums/prmun/}
}
\usage{
prmun
}
\description{
Nomes e códigos dos municípios do Paraná
}
\keyword{datasets}
